# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin panel API",
        "description": "Codeyad-React Learning-Ecommerce admin panel API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products",
                "description": "get all products or with pagination",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "[{}, {}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add product",
                "description": "store one product",
                "operationId": "addProduct",
                "requestBody": {
                    "description": "add one category",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "product test"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": "1500"
                                    },
                                    "weight": {
                                        "type": "string",
                                        "example": "2"
                                    },
                                    "brand_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "Description test for this product..."
                                    },
                                    "short_descriptions": {
                                        "type": "string",
                                        "example": "Description test for this product..."
                                    },
                                    "cart_descriptions": {
                                        "type": "string",
                                        "example": "Description test for this product when user selcted it in cart..."
                                    },
                                    "image": {
                                        "description": "file to upload",
                                        "type": "file"
                                    },
                                    "alt_image": {
                                        "type": "string",
                                        "example": "Some keyword"
                                    },
                                    "keywords": {
                                        "type": "string",
                                        "example": "keyword1-keyword2-keyword3"
                                    },
                                    "stock": {
                                        "type": "number",
                                        "example": "keyword1-keyword2-keyword3"
                                    },
                                    "discount": {
                                        "type": "number",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "[{}, {}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}